pull medium selector into main form
